Moved Objects implementations to another project.
