+++ Source/CodeMagic.Core/Items/IUsableItem.cs
@@ -1,0 +1,9 @@
+﻿using CodeMagic.Core.Game;
+
+namespace CodeMagic.Core.Items
+{
+    public interface IUsableItem
+    {
+        bool Use(IGameCore game);
+    }
+}
