+++ Source/CodeMagic.Implementations/IImageProvider.cs
@@ -1,0 +1,9 @@
+﻿using CodeMagic.UI.Images;
+
+namespace CodeMagic.Implementations
+{
+    public interface IImageProvider
+    {
+        SymbolsImage GetImage(IImagesStorage storage);
+    }
+}
