+++ Source/CodeMagic.Implementations/IImagesStorage.cs
@@ -1,0 +1,11 @@
+﻿using CodeMagic.UI.Images;
+
+namespace CodeMagic.Implementations
+{
+    public interface IImagesStorage
+    {
+        SymbolsImage GetImage(string name);
+
+        SymbolsImage[] GetAnimation(string name);
+    }
+}
