--- Source/CodeMagic.Objects.Implementation/CodeMagic.Objects.Implementation.csproj
@@ -1,76 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{BE94A033-ED03-4142-B634-563A110DD8BB}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>CodeMagic.Objects.Implementation</RootNamespace>
-    <AssemblyName>CodeMagic.Objects.Implementation</AssemblyName>
-    <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <Deterministic>true</Deterministic>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="AnimationManager.cs" />
-    <Compile Include="AnimationsBatchManager.cs" />
-    <Compile Include="CodeSpellImpl.cs" />
-    <Compile Include="Creatures\NonPlayable\GoblinImpl.cs" />
-    <Compile Include="DecorativeObjects\DecorativeObjectImpl.cs" />
-    <Compile Include="DecorativeObjects\FireObjectImpl.cs" />
-    <Compile Include="IceObjects\AcidIceImpl.cs" />
-    <Compile Include="IceObjects\WaterIceImpl.cs" />
-    <Compile Include="IImageProvider.cs" />
-    <Compile Include="Creatures\PlayerImpl.cs" />
-    <Compile Include="IImagesStorage.cs" />
-    <Compile Include="LiquidObjects\AcidLiquidImpl.cs" />
-    <Compile Include="LiquidObjects\OilLiquidImpl.cs" />
-    <Compile Include="LiquidObjects\WaterLiquidImpl.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SolidObjects\EnergyWallImpl.cs" />
-    <Compile Include="SolidObjects\WallImpl.cs" />
-    <Compile Include="SolidObjects\TorchWallImpl.cs" />
-    <Compile Include="SteamObjects\AcidSteamImpl.cs" />
-    <Compile Include="SteamObjects\WaterSteamImpl.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\CodeMagic.Core\CodeMagic.Core.csproj">
-      <Project>{5755a604-1bd8-4863-b0c7-4a15e5e1a6aa}</Project>
-      <Name>CodeMagic.Core</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\CodeMagic.UI.Images\CodeMagic.UI.Images.csproj">
-      <Project>{c5a8ee43-bc81-4e88-b990-d2ae962c3205}</Project>
-      <Name>CodeMagic.UI.Images</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-</Project>
