--- Source/CodeMagic.Objects.Implementation/IImageProvider.cs
@@ -1,9 +1,0 @@
-﻿using CodeMagic.UI.Images;
-
-namespace CodeMagic.Objects.Implementation
-{
-    public interface IImageProvider
-    {
-        SymbolsImage GetImage(IImagesStorage storage);
-    }
-}
