--- Source/CodeMagic.Objects.Implementation/IImagesStorage.cs
@@ -1,11 +1,0 @@
-﻿using CodeMagic.UI.Images;
-
-namespace CodeMagic.Objects.Implementation
-{
-    public interface IImagesStorage
-    {
-        SymbolsImage GetImage(string name);
-
-        SymbolsImage[] GetAnimation(string name);
-    }
-}
