+++ Source/CodeMagic.UI.Sad/Drawing/ImagesStorage.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text.RegularExpressions;
-using CodeMagic.Objects.Implementation;
+using CodeMagic.Implementations;
 using CodeMagic.UI.Images;
 
 namespace CodeMagic.UI.Sad.Drawing
