+++ Source/CodeMagic.sln
@@ -15,11 +15,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeMagic.UI.Sad", "CodeMag
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeMagic.UI.Images", "CodeMagic.UI.Images\CodeMagic.UI.Images.csproj", "{C5A8EE43-BC81-4E88-B990-D2AE962C3205}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeMagic.Objects.Implementation", "CodeMagic.Objects.Implementation\CodeMagic.Objects.Implementation.csproj", "{BE94A033-ED03-4142-B634-563A110DD8BB}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeMagic.UI.Sad.Common", "CodeMagic.UI.Sad.Common\CodeMagic.UI.Sad.Common.csproj", "{30048299-507D-4591-B6CA-18D247AE5CF1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeMagic.MapGeneration", "CodeMagic.MapGeneration\CodeMagic.MapGeneration.csproj", "{A4F02790-B78A-42D2-B1B2-BED6547E7FC3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeMagic.Implementations", "CodeMagic.Implementations\CodeMagic.Implementations.csproj", "{47E2B226-C00A-4EF3-AC0B-0A8A243ADD79}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -47,10 +47,6 @@ Global
 		{C5A8EE43-BC81-4E88-B990-D2AE962C3205}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C5A8EE43-BC81-4E88-B990-D2AE962C3205}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C5A8EE43-BC81-4E88-B990-D2AE962C3205}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BE94A033-ED03-4142-B634-563A110DD8BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BE94A033-ED03-4142-B634-563A110DD8BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BE94A033-ED03-4142-B634-563A110DD8BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BE94A033-ED03-4142-B634-563A110DD8BB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{30048299-507D-4591-B6CA-18D247AE5CF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{30048299-507D-4591-B6CA-18D247AE5CF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{30048299-507D-4591-B6CA-18D247AE5CF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -59,6 +55,10 @@ Global
 		{A4F02790-B78A-42D2-B1B2-BED6547E7FC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4F02790-B78A-42D2-B1B2-BED6547E7FC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A4F02790-B78A-42D2-B1B2-BED6547E7FC3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47E2B226-C00A-4EF3-AC0B-0A8A243ADD79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47E2B226-C00A-4EF3-AC0B-0A8A243ADD79}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47E2B226-C00A-4EF3-AC0B-0A8A243ADD79}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47E2B226-C00A-4EF3-AC0B-0A8A243ADD79}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
